requirements.txt
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
rich>=13.0.0
requests>=2.28.0
fredapi>=0.5.0
python-dotenv>=1.0.0
